Cleanup coordinator code
